moved useUpdateQueryParameter into separate hook
